requirements.txt
rich==13.7.0
typer==0.9.0
python-dotenv==1.0.0
openai==1.12.0 